backend/requirements.txt
sqlalchemy>=1.4.0
passlib[bcrypt]==1.7.4
email-validator>=1.1.0
python-dotenv>=0.19.0 
fastapi
uvicorn==0.27.1
pydantic
openai
httpx==0.24.1
pandas==2.1.3
python-jose==3.3.0
PyJWT==2.8.0
python-multipart==0.0.6
typing-extensions==4.9.0
psycopg2-binary==2.9.9
scipy
matplotlib
numpy
scikit-learn
scikit-image
seaborn
together==1.5.5